requirements.txt
python-telegram-bot==21.3
SQLAlchemy==2.0.36
psycopg==3.2.12
APScheduler==3.10.4
Flask==3.0.3